Clean up reuse lines in .md files.
We're battling duelling standards. REUSE wants the comments at the start
of a file and Doxygen wants the main title at the start of the file
otherwise it uses the file name. In the sole case of a one-word title
I've renamed the file. In the other cases I've created .license files.

In .md files which aren't used in the Doxygen docs, the REUSE lines
have been turned into comments so they don't appear in the output docs.