Validate BasisU transcoding.
In order to keep supporting stdin, this now uses std::istream for input
which can be that from a std::stringstream. If seek fails on stdin the
entir file is read into a stringstream. A streambufStream class that
wraps the std::streambuf in a ktxlib strea has been added.

Fixes #374.