Fix newly emerged warning from clang (#608)
and a warning from a recent code change in toktx.cc.

The new clang warning has appeared, without a bump in the LLVM version number, because Emscripten releases pull the top of tree LLVM and our CI gets the latest Emscripten release. LLVM is in the final throes of developing their 15.0.0 release. 

As there is no version number change we've had to use the `__has_warning` pre-processor directive and thus modify the offending BasisU file.

Note that MS VC++ also has `__has_warning` but with a different syntax so this fix uses the directive only when `__clang__` is defined.