Allow A8B8G8R8 formats. (#861)
Fixes issue #860.

Updates the switches used by the the GL loader to those generated by
`generate_format_switches.rb` in the KTX-Specification repo as a way to
get them updated for these and some other recently added Vulkan formats.
The glloader now loads the newly allowed formats.

VkFormat declarations for Java and Python are now generated by
`mkvkformatfiles` which has been moved to ci_scripts since it now
generates files outside the `lib` directory. The unattractive alternative
was to hand edit these files to add the newly allowed formats.

CI testing of vkformat-related file generation (`mkvk`) has been moved
to a dedicated GitHub workflow and removed from the main CI builds. This
is because testing now requires a clone of the KTX-Specification repo due
to the use of `generate_format_switches.rb`.

The change to `createdfd.c` is to remove a duplicated line of code and
is unrelated to the purpose of this PR.

https://github.com/KhronosGroup/KTX-Software-CTS/pull/19 removes the
tests for these as prohibited and adds new tests for the formats.