Properly convert Java char to C char in inputSwizzle (#876)
Fixes https://github.com/KhronosGroup/KTX-Software/issues/875 

As detailed in the issue: The Java `char[]` array for the input swizzle
was casted to a `jbyteArray` in the JNI layer, in order to write it to
the C/C++ `char[4]` array. This cast was invalid and caused data
corruption.

There could be two ways of tackling this:

- One could change the type of the `inputSwizzle` in the Java layer to
`byte[]` (which has 8 bits, and would make the cast valid). This would
be a breaking change, and may require the inconvenient casting at the
call site like
`p.setInputSwizzle(new byte[]{ (byte)'b', (byte)'r', (byte)'g',
(byte)'a' });`
- One could fetch the Java `char[]` array as a `jcharArray`, and cast
the elements to (C/C++) `char` individually

This PR takes the second option (but we could do this either way,
depending on the preferences from others).