Wrapper script to use ktx compare for git diff.
Those wishing to use this must run, or have run, install-gitconfig.{ps1,sh}
so that .gitconfig, which now includes a "ktx-compare" diff command will
be included by your repo clone's `.git/config`.

They need to have the `ktx` command installed and in the $PATH.

They need to add the line

  *.ktx2 binary diff=ktx-compare

to their clone's .git/info/attributes or to a global Git attributes file.
This is not included in the repo's .gitattributes because not everyone
will have the ktx command installed.