docs: clarify pointer lifetimes for ktxTexture_SetImageFromMemory
ktxTexture1_SetImageFromMemory and ktxTexture2_SetImageFromMemory accept
a pointer to image memory but do not clearly document if ownership of
those pointers is transferred or copied. This commit adds a doxygen
warning informing users that the underlying image data is copied and
therefore they are responsible for their copy of the image memory.

Closes #914