Workaround pyparser install failure on Travis Ubuntu/arm64. (#924)
Unbelievable. While still wrangling with the workaround for the major
GHA Windows runner image meltdown (PR #920 and blocked-by-this PR #923)
Travis breaks their Ubuntu/arm64 runner image such that installs of
python-related tools on to the VM are failing.

This is their suggested workaround. I don't yet know if it will be
needed permanently.