Update load tests to Vulkan SDK version to 1.3.290.
This has first release of official iOS support and first Windows arm64 SDK.

Stop using binaries of SDL2, assimp and glew libraries stored in repo
because an updated SDL2 is needed for 1.3.290 and more platforms are
now supported. Dependencies for iOS, macOS and Windows are now obtained
through vcpkg.

These changes plus removal of the hacks for finding the iOS in older
Vulkan SDKs means a major revamp of the load tests-related CMake files.