Java swizzle and constant fixes (#972)
Fixes https://github.com/KhronosGroup/KTX-Software/issues/969 (see the
issue for details).

The fix for the `inputSwizzle` (for both the `KtxAstcParams` and
`KtxBasisParams`) is implemented by checking whether the input equals
the default, which is a `char[] { 0, 0, 0, 0 }` array. If this is the
case, then this value will be accepted, passed to the native layer, and
handled there accordingly. Only when the input is _not_ that default,
the elements will be checked to be in `rgba01`.