Hide the ASTCENC_DECOMPRESSOR option and set it to OFF
It's an option provided by external/astc-encoder/CMakeLists.txt
but setting it to ON in KTX-Software breaks the build.
So set it to OFF and make it an INTERNAL variable so it's not shown
in cmake-gui where it could lure users to enable it.

While at it, also fixed the line number in a comment,
and a typo in another comment.