Fix for Emscripten 4.0.9 (#1026)
Includes the following fixes to work with Emscripten 4.0.9:

- Install sdl2 with `embuilder` as SDL2's config file is no longer
  installed in SYSROOT without this explicit step.
- Explicitly allow_raw_pointers() when returning TranscodedImage.
  Emscripten 4.0.8+ require that this policy be specified.
- Add HEAP8 method to EXPORTED_RUNTIME_METHODS. Emscripten 4.0.9
  removes this method, which is used in the binding.

Fix `scripts/build_wasm_docker.sh` to not remove the emscripten Docker
container when done so the next run will benefit from the Emscripten
cache. Removal can be requested with `-r`.

Fix handling of the KTX_PACK_ASTC_ENCODER_MODE_DEFAULT. When
passed in params.mode to ktxTexture2_EncodeAstcEx this value triggered
an assert. The function now picks a  mode appropriate to the bit-depth of
the input texture. Document this handling in ktx.h.