Export ktxTexture2_DecodeAstc in JS bindings (#1034)
Add ktxTexture2_DecodeAstc to `libktx_read`, a previous omission. Needed
too so the function can be included in the read-only JS bindings.

Add tests of ASTC encode and decode functionality to `texturetest`. The
change includes needing to pass the path to `ktxdiff` to `texturetests`
as the new tests use it to compare the decoded result with the original
image.

Change `texturetest` to use std::filesystem::path and c++17.

Fix issue in `astc_codec.cpp` where conversion of `params.inputSwizzle'
could fail, asserting in a debug build.